+++ src/utils/bot/DanielBotClasses/moves.java
@@ -2,11 +2,11 @@ package utils.bot.DanielBotClasses;
 
 // movelist class. it's literally an arraylist, but since arraylists are slow, we have a count index
 public class moves {
-	public int moves[];
+	public short moves[];
 	public int count;
 	
 	public moves() {
-		moves = new int[256];
+		moves = new short[256];
 		count = 0;
 	}
 }
